% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serosurvey_viz.R
\name{ggplot_prevalence}
\alias{ggplot_prevalence}
\alias{ggplot_prevalence_ii}
\title{Visualization of proportions}
\usage{
ggplot_prevalence(data, category, outcome, proportion, proportion_upp,
  proportion_low, breaks_n = 5)

ggplot_prevalence_ii(data, denominator_level, numerator, proportion,
  proportion_upp, proportion_low, breaks_n = 5)
}
\arguments{
\item{data}{input tibble}

\item{category}{denominator level}

\item{outcome}{numerator variable}

\item{proportion}{point estimate}

\item{proportion_upp}{upper interval}

\item{proportion_low}{lower interval}

\item{breaks_n}{number of breaks in axis}

\item{denominator_level}{\itemize{
\item denominator values column
}}

\item{numerator}{\itemize{
\item numerator variable name column
}}
}
\description{
visualization of proportions
}
\section{Functions}{
\itemize{
\item \code{ggplot_prevalence}: ggplot2 visualization of proportions

\item \code{ggplot_prevalence_ii}: ggplot_prevalence with new arguments
}}

\examples{

# not yet
# go to workflow:
# https://avallecam.github.io/serosurvey/articles/howto-reprex.html

}
