% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serosurvey_misc.R
\name{unite_dotwhiskers}
\alias{unite_dotwhiskers}
\alias{serosvy_extract_posterior}
\title{Miscellaneous function}
\usage{
unite_dotwhiskers(
  data,
  variable_dot,
  variable_low,
  variable_upp,
  digits_dot = 3,
  digits_low = 2,
  digits_upp = 3,
  decimal_to_percent = TRUE
)

serosvy_extract_posterior(data, variable)
}
\arguments{
\item{data}{tibble or dataframe}

\item{variable_dot}{point estimate}

\item{variable_low}{lower interval}

\item{variable_upp}{upper interval}

\item{digits_dot}{digits of point estimate to round}

\item{digits_low}{digits of lower interval to round}

\item{digits_upp}{digits of upper interval to round}

\item{decimal_to_percent}{logical to multiply values by 100. TRUE by default.}

\item{variable}{variable to extract from posterior distributions}
}
\description{
Miscellaneous function
}
\section{Functions}{
\itemize{
\item \code{unite_dotwhiskers}: create new columns with unified point estimate and intervals with round numbers

\item \code{serosvy_extract_posterior}: priorización con dos covariables
}}

\examples{

\dontrun{

library(tidyverse)

test <- rgamma(n = 1000,shape = 3,rate = 0.2) \%>\%
  enframe(name = NULL) \%>\%
  mutate(value=value/100) \%>\%
  skimr::skim() \%>\%
  as_tibble() \%>\%
  select(contains("mean"),contains("p25"),contains("p50"))

test \%>\% glimpse()

test \%>\%
  unite_dotwhiskers(variable_dot = numeric.mean,
                    variable_low = numeric.p25,
                    variable_upp = numeric.p50,
                    digits_dot = 3,
                    digits_low = 3,
                    digits_upp = 3)
}

}
