% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serosurvey_viz.R
\name{ggplot_prevalence}
\alias{ggplot_prevalence}
\title{Visualization of proportions}
\usage{
ggplot_prevalence(
  data,
  denominator_level,
  numerator,
  proportion,
  proportion_upp,
  proportion_low
)
}
\arguments{
\item{data}{input tibble}

\item{denominator_level}{\itemize{
\item denominator values column
}}

\item{numerator}{\itemize{
\item numerator variable name column
}}

\item{proportion}{point estimate}

\item{proportion_upp}{upper interval}

\item{proportion_low}{lower interval
#param breaks_n number of breaks in axis}
}
\description{
visualization of proportions
}
\section{Functions}{
\itemize{
\item \code{ggplot_prevalence}: ggplot2 visualization of proportions
}}

\examples{

# not yet
# go to workflow:
# https://avallecam.github.io/serosurvey/articles/howto-reprex.html

}
